+++ hir/src/ops/array/squeeze.rs
@@ -50,7 +50,7 @@ impl Expansion for Squeeze {
         check_output_arity(&outputs, 1)?;
         s.equals(&outputs[0].datum_type, &inputs[0].datum_type)?;
         if let Some(ref axes) = self.axes {
-            s.equals(&outputs[0].rank, (&inputs[0].rank).bex() - axes.len() as i32)?;
+            s.equals(&outputs[0].rank, (&inputs[0].rank).bex() - axes.len() as i64)?;
         }
         s.given(&inputs[0].shape, move |s, shape| {
             let output_shape = self.compute_shape(&shape)?;
