+++ hir/src/ops/nn/reduce.rs
@@ -192,7 +192,7 @@ impl Expansion for Reduce {
         if self.keep_dims {
             s.equals(&inputs[0].rank, &outputs[0].rank)?;
         } else if let Some(axes) = self.axes.as_ref() {
-            s.equals(inputs[0].rank.bex() - axes.len() as i32, &outputs[0].rank)?;
+            s.equals(inputs[0].rank.bex() - axes.len() as i64, &outputs[0].rank)?;
         } else {
             s.equals(&outputs[0].rank, 0)?;
         }
